+++ lib/src/widgets/eyedrop/eyedropper_button_web.dart
@@ -1,3 +1,4 @@
+// ignore: avoid_web_libraries_in_flutter
 import 'dart:js' as js;
 
 import 'package:flutter/material.dart';
@@ -16,11 +17,15 @@ class EyedropperButton extends StatelessWidget {
   /// color selection callback
   final ValueChanged<Color> onColor;
 
+  /// hover, and the color changed callback
+  final ValueChanged<Color>? onColorChanged;
+
   /// verify if the button is in a CanvasKit context
   bool get eyedropEnabled => js.context['flutterCanvasKit'] != null;
 
   const EyedropperButton({
     required this.onColor,
+    this.onColorChanged,
     this.icon = Icons.colorize,
     this.iconColor = Colors.blueGrey,
     Key? key,
@@ -41,7 +46,7 @@ class EyedropperButton extends StatelessWidget {
 
   void _onEyeDropperRequest(BuildContext context) {
     try {
-      EyeDrop.of(context).capture(context, onColor);
+      EyeDrop.of(context).capture(context, onColor, onColorChanged);
     } catch (err) {
       throw Exception('EyeDrop capture error : $err');
     }