+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-# 0.5.2
+# 0.6.1
 
-- Update the colorpicker selectedColor "internally" to simplify the usage without a ColorButton
+- Update the deprecated theme properties
+- Migrate to Flutter 3.7 slider API : Thx [@alonohana627](https://github.com/alonohana627)
+
+# 0.6.0
+
+- add a `onColorChanged` callback to send the hovered color during eyedropping. Thx [@talisk](https://github.com/talisk)
 
 # 0.5.1